requirements.txt
awscrt==0.31.3
awsiotsdk==1.28.2
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==2.1.5
packaging==26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.5
six==1.17.0
urllib3==1.26.20
Werkzeug==3.0.6
zipp==3.20.2
